Fix an issue where on password reset/account activation, the JS code to update password would be called. The remedy was to restore the <form> inside the account page for update password and move the js code from func.mako to index.mako in order to make the code less confusing.
The changes create a bug where a password update will not show feedback, because POSTing the update password code reloads the index.mako template, not allowing the data-alert to show in time. A possible fix to this would be to fix the @view resetpassword such that it will handle the password update POST and show the success/fail message accordingly.

Also fix a celery warning about using pickle by updating the celery config params to accept ['pickle', 'json', 'msgpack', 'yaml’].